Change the 'server' and 'build' makefile targets
These will now use the Dockerfile, building the
image if necessary (using a .built-docker-image
file to keep track of dependencies, in case the
Dockerfile or Gemfile are modified).

Once built, the `middleman server` or `middleman 
build` commands are invoked, with the `source` and
`docs` directories mounted into the docker image.