Tell CircleCI to build from the master branch
I thought we might have to merge into a different
branch, to avoid an endless build/push/build/...
cycle. But, we can do that using the shell script
that detects if the last commit was made via the
build pipeline. So, we can keep things simple and
build from the master branch, as normal.