Mount local config directory when serving/building
Prior to this change, the config settings used by middleman will
be whatever was in place when the docker image was first built.

So, any subsequent config changes will have no effect.
This change ensures that the latest config settings are used,
whenever the middleman server or build commands are invoked.

The CircleCI build pipeline would not be affected by this
problem, because it is always working from a full checkout
of the source, including whatever is in the config directory.
So, without this change, config changes would have affected the
live version of the user guide, but would not have been visible to
the developer, using the docker image for local development.