Move guide restructuring section near to the top
Having the temporary 'this guide is being restructured' section near the
top of the page means it is visible on initial page load, when using a
browser with default page zoom and font size. So, it is more likely that
people will read it, and understand the new structure with less
confusion.