Update custom domain documentation
Since the original version of this document, we have refined the process. We will be managing all custom zones in
our cloud-platform account and so regardless of the existing setup, we ask users to define the route53 zone as a
resource.

Managing specific cases (migrating domains, third-party administrators) can be done at the time when the NS records
need to be setup, to delegate control of the zone to our account.

This should simplify the overall process by introducing a single gate: setting up the NS records.