Stop building the docker image for 'make server'
Most of the time, users will not need to rebuild the docker image,
when writing/modifying articles in the user guide. So, rebuilding
it for every user is a waste of time.

This commit tells the makefile to pull the docker image, rather
than building it, for the 'make server' command.