Update 'creating namespace' yaml files
We have changed the yaml file templates we use to
create new namespaces. This commit updates the
user guide with the new files, and removes the
reference to 'hard limits' in the resource quota
section (because we no longer use hard limits).

Related to: https://github.com/ministryofjustice/cloud-platform-environments/pull/1038