Restructure and simplify the namespace article
This article went into a lot of detail from the
POV of the cloud platform. This commit reduces the
complexity we present to the user, and hopefully
makes it easier to follow the steps.

We have a script which prompts the user for the
information required, and creates the relevant
files in the correct path in the environments
repo. So, we don't need a lot of information here
about what values are needed, and how the repo is
structured.

Keeping all the information in the prompts
provided by the script makes it easier to keep
them up to date.

The user no longer needs to install terraform,
and there is only one cluster.