Bump the image version
This commit is mainly here as a record that I pushed a new version
of the docker image to docker hub.

After #135, the circleci build pipeline was broken, but we didn't
notice because there haven't been any content updates since then.
Updating the image on docker hub fixed the circleci build pipeline