Change the service.yaml file to use 'https'
I went through a very similar process to this, yesterday, and the
app I deployed was not accessible via the URL until I changed the
port name in the spec. from 'http' to 'https'.

I think we must have changed something about the way in which
ingresses/services work, by default, since the article was
originally written.

I haven't worked through this example, since making the change,
but I'm confident this is an issue, based on this:

https://github.com/ministryofjustice/cloud-platform-environments/pull/1181