Add a section on "OFFICIAL SENSITIVE" services
We sometimes get asked whether or not we can host services in the
cloud platform which handle "OFFICIAL SENSITIVE" data.

This change adds an article to the user guide which attempts to
clarify our position on this.

I've chatted with the folks in #security, and there isn't a formal
assurance document we can point to which gives a blanket OK for
hosting on the cloud platform, so this is about as definite as we
can be, for now.