Remove unnecessary 'terraform' stanza
By the time users are ready to create a pingdom check, they probably
already have some AWS resources in their namespaces, and so the
terraform stanza will already be present.

Having it here will cause problems if the user copies and pastes this
code, so this change removes that stanza from the example.