Add ruby code to manage the publishing process
The 'should_rebuild?' logic here may not be
needed, since the github action will only trigger
when a PR is closed. Since publishing the site
pushes straight to the master branch, without
generating a PR, that should be enough to avoid
an infinite loop, even without this test. Still,
it doesn't do any harm, so I'm leaving it in.