Use force push to publish compiled pages
The branch protection on this repo will not allow the github action user
to push to the master branch. But, there is an option to allow
force-pushing by all users, which I have now enabled.

So, changing this script to force push the compiled changes should allow
the publishing process to work.