Update publishing process
* Remove `should_rebuild?` test

We don't need this, since we're not in danger of
getting into the same infinite loop as with 
circleci, because this action will trigger on a 
PR being opened, not on a change to the master
branch.

* Change the process to commit updated files

We're pushing to the PR branch, which is exactly
what the code-formatter github action does. So, we
can use the same process.