Remove compiled version of the migration guide
The publishing script throws an error when you delete a file. I think it
might require quite a lot of work to enable the script to do that, so in
this case I'm just going to manually delete the compiled HTML file.
Hopefully, this will enable the github action to complete successfully.