Add a github action to keep "master" in sync
We want to use "main" as the default branch, but we can only
publish the "docs" folder of the "master" branch. This is a limitation
of github pages.

This change adds a github action to automatically keep the "master"
branch in sync with the "main" branch. So although we will still have a
"master" branch until github fix the problem, at least the default
branch will be "main"