Use publisher docker image 1.1
This change should make no difference at all, since this github action does not set a "ROOT_DOCPATH" env. var (which is what the 1.1 image version uses to publish sites differently if they are not hosted at the root of a domain).

I'm making this change solely to keep the image versions consistent across our documentation sites, and to confirm that there really is no impact from making this change.