Add a note not to remove TD migration page.
This page is sometimes referenced for general migration advice. When we remove template deploy-specific pages, content on this page should be replaced by a generic migration advice page, and this page should link to that one, so that anyone following an old link can still find the information they need.