Add "About" page - product description
We should probably describe our product offering somewhere, to help people who are trying to understand the aims and values of Cloud Platform, whether they should select it and whether there are other options.

Is this User Guide the best place? It's pretty sharply "get the tech task done" focused currently, so maybe there is a better place? But the User Guide seems most like a home page for the platform.