Remove link to the new About page
related to #435

That PR added a new page, and a link to it on the index page, in a
single PR. This is a perfectly sensible thing to do, but unfortunately
it means that the link-checker fails because the home page has a link to
a page that the link-checker doesn't know exists yet.

This is a bug, and should be fixed, but in the meantime I'm raising 2
PRs to get this change deployed successfully - one to add the new page,
and then a second to add the link to the page. That should keep the
link-checker happy.