Add doc "What can be hosted on CP"
This is to be clear about expectations for apps that go on Cloud Platform.

Why? Because in our migration to EKS we are being clear that state should
not live in containers, and I think this must be the first time we've talked
about apps needing to be 12 Factor Apps. This should be part of the deal
when considering CP.

Also, Analytical Platform need to factor this into their thinking. And
Serverless is a recent topic that needs a bit of addressing.

Whilst doing this, I changed the index page a bit:
* "CP's offering" section, to house this new article. The 'About' page fits
  well here too. I think this stuff is important enough to be at the top of
  the index.
* Added comments to each section to start to think about who they are for and why,
  because the sections are a bit of a mess at the moment. This is to start
  the ball rolling.