Tweaks to guide to publish prototypes
Added the explicit Git commands to run to create a new branch, commit
the changes and push to GitHub. Some designers may be less familiar
with the command line and Git.

In addition to this, page elaborates a bit on how to create a PR.

Similarly, in the Cloud Platform CLI page I added a note to made it clear
brew needs to be installed on your machine to run `brew install ...`.
Again, this may not be immediately apparent to some users (the note
refers them to Homebrew website for instructions on how to install it).